+++ tests/trustless_gateway_car_test.go
@@ -11,9 +11,9 @@ import (
 )
 
 func TestTrustlessCarPathing(t *testing.T) {
-	subdirTwoSingleBlockFilesFixture := car.MustOpenUnixfsCar("t0118/subdir-with-two-single-block-files.car")
-	singleLayerHamtMultiBlockFilesFixture := car.MustOpenUnixfsCar("t0118/single-layer-hamt-with-multi-block-files.car")
-	dirWithDagCborWithLinksFixture := car.MustOpenUnixfsCar("t0118/dir-with-dag-cbor-with-links.car")
+	subdirTwoSingleBlockFilesFixture := car.MustOpenUnixfsCar("trustless_gateway_car/subdir-with-two-single-block-files.car")
+	singleLayerHamtMultiBlockFilesFixture := car.MustOpenUnixfsCar("trustless_gateway_car/single-layer-hamt-with-multi-block-files.car")
+	dirWithDagCborWithLinksFixture := car.MustOpenUnixfsCar("trustless_gateway_car/dir-with-dag-cbor-with-links.car")
 
 	tests := SugarTests{
 		{
@@ -59,7 +59,7 @@ func TestTrustlessCarPathing(t *testing.T) {
 							singleLayerHamtMultiBlockFilesFixture.MustGetCid(),
 							singleLayerHamtMultiBlockFilesFixture.MustGetCIDsInHAMTTraversal(nil, "685.txt"),
 							singleLayerHamtMultiBlockFilesFixture.MustGetCid("685.txt"),
-							singleLayerHamtMultiBlockFilesFixture.MustGetChildrenCids("685.txt"))...,
+							singleLayerHamtMultiBlockFilesFixture.MustGetDescendantsCids("685.txt"))...,
 						).
 						Exactly().
 						InThatOrder(),
@@ -94,8 +94,8 @@ func TestTrustlessCarPathing(t *testing.T) {
 }
 
 func TestTrustlessCarDagScopeBlock(t *testing.T) {
-	subdirTwoSingleBlockFilesFixture := car.MustOpenUnixfsCar("t0118/subdir-with-two-single-block-files.car")
-	singleLayerHamtMultiBlockFilesFixture := car.MustOpenUnixfsCar("t0118/single-layer-hamt-with-multi-block-files.car")
+	subdirTwoSingleBlockFilesFixture := car.MustOpenUnixfsCar("trustless_gateway_car/subdir-with-two-single-block-files.car")
+	singleLayerHamtMultiBlockFilesFixture := car.MustOpenUnixfsCar("trustless_gateway_car/single-layer-hamt-with-multi-block-files.car")
 
 	tests := SugarTests{
 		{
@@ -177,10 +177,10 @@ func TestTrustlessCarDagScopeBlock(t *testing.T) {
 }
 
 func TestTrustlessCarDagScopeEntity(t *testing.T) {
-	subdirTwoSingleBlockFilesFixture := car.MustOpenUnixfsCar("t0118/subdir-with-two-single-block-files.car")
-	singleLayerHamtMultiBlockFilesFixture := car.MustOpenUnixfsCar("t0118/single-layer-hamt-with-multi-block-files.car")
-	subdirWithMixedBlockFiles := car.MustOpenUnixfsCar("t0118/subdir-with-mixed-block-files.car")
-	dirWithDagCborWithLinksFixture := car.MustOpenUnixfsCar("t0118/dir-with-dag-cbor-with-links.car")
+	subdirTwoSingleBlockFilesFixture := car.MustOpenUnixfsCar("trustless_gateway_car/subdir-with-two-single-block-files.car")
+	singleLayerHamtMultiBlockFilesFixture := car.MustOpenUnixfsCar("trustless_gateway_car/single-layer-hamt-with-multi-block-files.car")
+	subdirWithMixedBlockFiles := car.MustOpenUnixfsCar("trustless_gateway_car/subdir-with-mixed-block-files.car")
+	dirWithDagCborWithLinksFixture := car.MustOpenUnixfsCar("trustless_gateway_car/dir-with-dag-cbor-with-links.car")
 
 	tests := SugarTests{
 		{
@@ -275,7 +275,7 @@ func TestTrustlessCarDagScopeEntity(t *testing.T) {
 								subdirWithMixedBlockFiles.MustGetCid(),
 								subdirWithMixedBlockFiles.MustGetCid("subdir"),
 								subdirWithMixedBlockFiles.MustGetCid("subdir", "multiblock.txt"),
-								subdirWithMixedBlockFiles.MustGetChildrenCids("subdir", "multiblock.txt"),
+								subdirWithMixedBlockFiles.MustGetDescendantsCids("subdir", "multiblock.txt"),
 							)...,
 						).
 						Exactly().
@@ -311,7 +311,7 @@ func TestTrustlessCarDagScopeEntity(t *testing.T) {
 }
 
 func TestTrustlessCarDagScopeAll(t *testing.T) {
-	subdirWithMixedBlockFiles := car.MustOpenUnixfsCar("t0118/subdir-with-mixed-block-files.car")
+	subdirWithMixedBlockFiles := car.MustOpenUnixfsCar("trustless_gateway_car/subdir-with-mixed-block-files.car")
 
 	tests := SugarTests{
 		{
@@ -334,7 +334,7 @@ func TestTrustlessCarDagScopeAll(t *testing.T) {
 								t,
 								subdirWithMixedBlockFiles.MustGetCid(),
 								subdirWithMixedBlockFiles.MustGetCid("subdir"),
-								subdirWithMixedBlockFiles.MustGetChildrenCids("subdir"),
+								subdirWithMixedBlockFiles.MustGetDescendantsCids("subdir"),
 							)...,
 						).
 						Exactly().
@@ -361,7 +361,7 @@ func TestTrustlessCarDagScopeAll(t *testing.T) {
 								subdirWithMixedBlockFiles.MustGetCid(),
 								subdirWithMixedBlockFiles.MustGetCid("subdir"),
 								subdirWithMixedBlockFiles.MustGetCid("subdir", "multiblock.txt"),
-								subdirWithMixedBlockFiles.MustGetChildrenCids("subdir", "multiblock.txt"),
+								subdirWithMixedBlockFiles.MustGetDescendantsCids("subdir", "multiblock.txt"),
 							)...,
 						).
 						Exactly().
@@ -374,10 +374,60 @@ func TestTrustlessCarDagScopeAll(t *testing.T) {
 }
 
 func TestTrustlessCarEntityBytes(t *testing.T) {
-	singleLayerHamtMultiBlockFilesFixture := car.MustOpenUnixfsCar("t0118/single-layer-hamt-with-multi-block-files.car")
-	subdirWithMixedBlockFiles := car.MustOpenUnixfsCar("t0118/subdir-with-mixed-block-files.car")
+	singleLayerHamtMultiBlockFilesFixture := car.MustOpenUnixfsCar("trustless_gateway_car/single-layer-hamt-with-multi-block-files.car")
+	subdirWithMixedBlockFiles := car.MustOpenUnixfsCar("trustless_gateway_car/subdir-with-mixed-block-files.car")
+	missingBlockFixture := car.MustOpenUnixfsCar("trustless_gateway_car/file-3k-and-3-blocks-missing-block.car")
 
 	tests := SugarTests{
+		{
+			Name: "GET CAR with entity-bytes succeeds even if the gateway is missing a block after the requested range",
+			Hint: `
+				dag-scope=entity&entity-bytes=0:x should return a CAR file with
+				only the blocks needed to fullfill the request. This MUST
+				succeed despite the fact that bytes beyond 'x' are not retrievable.
+			`,
+			Request: Request().
+				Path("/ipfs/{{cid}}", missingBlockFixture.MustGetCidWithCodec(0x70)).
+				Query("format", "car").
+				Query("dag-scope", "entity").
+				Query("entity-bytes", "0:1000"),
+			Response: Expect().
+				Status(200).
+				Body(
+					IsCar().
+						IgnoreRoots().
+						HasBlocks(
+							missingBlockFixture.MustGetCid(),
+							missingBlockFixture.MustGetChildrenCids()[0],
+						).
+						Exactly(),
+				),
+		},
+		{
+			Name: "GET CAR with entity-bytes succeeds even if the gateway is missing a block before the requested range",
+			Hint: `
+				dag-scope=entity&entity-bytes=y:* should return a CAR file with
+				only the blocks needed to fullfill the request. This MUST
+				succeed despite the fact that a block with bytes before 'y' is
+				not retrievable.
+			`,
+			Request: Request().
+				Path("/ipfs/{{cid}}", missingBlockFixture.MustGetCidWithCodec(0x70)).
+				Query("format", "car").
+				Query("dag-scope", "entity").
+				Query("entity-bytes", "2200:*"),
+			Response: Expect().
+				Status(200).
+				Body(
+					IsCar().
+						IgnoreRoots().
+						HasBlocks(
+							missingBlockFixture.MustGetCid(),
+							missingBlockFixture.MustGetChildrenCids()[2],
+						).
+						Exactly(),
+				),
+		},
 		{
 			Name: "GET CAR with entity-bytes of a full UnixFS file",
 			Hint: `
@@ -398,7 +448,7 @@ func TestTrustlessCarEntityBytes(t *testing.T) {
 							subdirWithMixedBlockFiles.MustGetCid(),
 							subdirWithMixedBlockFiles.MustGetCid("subdir"),
 							subdirWithMixedBlockFiles.MustGetCid("subdir", "multiblock.txt"),
-							subdirWithMixedBlockFiles.MustGetChildrenCids("subdir", "multiblock.txt"),
+							subdirWithMixedBlockFiles.MustGetDescendantsCids("subdir", "multiblock.txt"),
 						)...).
 						Exactly().
 						InThatOrder(),
@@ -448,7 +498,7 @@ func TestTrustlessCarEntityBytes(t *testing.T) {
 						HasBlocks(
 							flattenStrings(t,
 								subdirWithMixedBlockFiles.MustGetCid("subdir", "multiblock.txt"),
-								subdirWithMixedBlockFiles.MustGetChildrenCids("subdir", "multiblock.txt")[2:])...,
+								subdirWithMixedBlockFiles.MustGetDescendantsCids("subdir", "multiblock.txt")[2:])...,
 						).
 						Exactly().
 						InThatOrder(),
@@ -472,7 +522,7 @@ func TestTrustlessCarEntityBytes(t *testing.T) {
 						HasBlocks(
 							flattenStrings(t,
 								subdirWithMixedBlockFiles.MustGetCid("subdir", "multiblock.txt"),
-								subdirWithMixedBlockFiles.MustGetChildrenCids("subdir", "multiblock.txt")[2:4])...,
+								subdirWithMixedBlockFiles.MustGetDescendantsCids("subdir", "multiblock.txt")[2:4])...,
 						).
 						Exactly().
 						InThatOrder(),
@@ -496,7 +546,7 @@ func TestTrustlessCarEntityBytes(t *testing.T) {
 						HasBlocks(
 							flattenStrings(t,
 								subdirWithMixedBlockFiles.MustGetCid("subdir", "multiblock.txt"),
-								subdirWithMixedBlockFiles.MustGetChildrenCids("subdir", "multiblock.txt")[2:4])...,
+								subdirWithMixedBlockFiles.MustGetDescendantsCids("subdir", "multiblock.txt")[2:4])...,
 						).
 						Exactly().
 						InThatOrder(),
@@ -520,7 +570,7 @@ func TestTrustlessCarEntityBytes(t *testing.T) {
 						HasBlocks(
 							flattenStrings(t,
 								subdirWithMixedBlockFiles.MustGetCid("subdir", "multiblock.txt"),
-								subdirWithMixedBlockFiles.MustGetChildrenCids("subdir", "multiblock.txt")[3:])...,
+								subdirWithMixedBlockFiles.MustGetDescendantsCids("subdir", "multiblock.txt")[3:])...,
 						).
 						Exactly().
 						InThatOrder(),
@@ -544,7 +594,7 @@ func TestTrustlessCarEntityBytes(t *testing.T) {
 						HasBlocks(
 							flattenStrings(t,
 								subdirWithMixedBlockFiles.MustGetCid("subdir", "multiblock.txt"),
-								subdirWithMixedBlockFiles.MustGetChildrenCids("subdir", "multiblock.txt")[:5])...,
+								subdirWithMixedBlockFiles.MustGetDescendantsCids("subdir", "multiblock.txt")[:5])...,
 						).
 						Exactly().
 						InThatOrder(),
@@ -568,7 +618,7 @@ func TestTrustlessCarEntityBytes(t *testing.T) {
 						HasBlocks(
 							flattenStrings(t,
 								subdirWithMixedBlockFiles.MustGetCid("subdir", "multiblock.txt"),
-								subdirWithMixedBlockFiles.MustGetChildrenCids("subdir", "multiblock.txt")[0])...,
+								subdirWithMixedBlockFiles.MustGetDescendantsCids("subdir", "multiblock.txt")[0])...,
 						).
 						Exactly().
 						InThatOrder(),