+++ tooling/car/fixture_test.go
@@ -21,7 +21,7 @@ func TestGetNodes(t *testing.T) {
 	leaf := f.MustGetNode("subdir", "leaf.txt").Cid().String()
 	assert.Equal(t, "bafkreihdhgb5vyuqu7jssreyo3h567obewtqq37fi5hr2w4um5icacry7m", leaf)
 
-	nodes := f.MustGetChildren()
+	nodes := f.MustGetDescendants()
 
 	assert.Len(t, nodes, 4)
 	assert.Equal(t, "bafkreidw23elffhagxz3oi6ctoibqouzfowfn3bwcvq2yzgd5n5h4gjyou", nodes[0].Cid().String())
@@ -29,7 +29,7 @@ func TestGetNodes(t *testing.T) {
 	assert.Equal(t, "bafybeiaq6e55xratife7s5cmzjcmwy4adzzlk74sbdpfcq72gus6cweeeq", nodes[2].Cid().String())
 	assert.Equal(t, "bafkreihdhgb5vyuqu7jssreyo3h567obewtqq37fi5hr2w4um5icacry7m", nodes[3].Cid().String())
 
-	cids := f.MustGetChildrenCids()
+	cids := f.MustGetDescendantsCids()
 	assert.Len(t, nodes, 4)
 	assert.Equal(t, "bafkreidw23elffhagxz3oi6ctoibqouzfowfn3bwcvq2yzgd5n5h4gjyou", cids[0])
 	assert.Equal(t, "bafkreiaeqsxxqwmsnhzhrlyr2udn25hpj24bs7gzcgkhbrkmhcuikcgh4a", cids[1])
@@ -46,7 +46,7 @@ func TestIssue54MustGetChildren(t *testing.T) {
 	// QmZgfvZtoFdbJy4JmpPHc1NCXyA7Snim2L8e6zKspiUzhu  7       ./sub1/hello.txt
 	// QmVtAZGRHTCzSNt1vRgz1UESvcc57ebEcDYTaDJjVu1SrA  -       ./sub2
 	// Qmf4EqZZpFPcy6oKsc84dGS5EpPdXYZ1hq39Gemadu6hfW  7       ./sub2/hello.txt
-	cids := f.MustGetChildrenCids("sub1", "hello.txt")
+	cids := f.MustGetDescendantsCids("sub1", "hello.txt")
 
 	// › ipfs dag get QmZgfvZtoFdbJy4JmpPHc1NCXyA7Snim2L8e6zKspiUzhu | jq
 	// {
@@ -93,9 +93,54 @@ func TestIssue54MustGetChildren(t *testing.T) {
 	//     }
 	//   ]
 	// }
-	cids = f.MustGetChildrenCids("sub1")
+	cids = f.MustGetDescendantsCids("sub1")
 	assert.Len(t, cids, 3)
 	assert.Equal(t, "QmZgfvZtoFdbJy4JmpPHc1NCXyA7Snim2L8e6zKspiUzhu", cids[0])
 	assert.Equal(t, "QmaATBg1yioWhYHhoA8XSUqD1Ya91KKCibWVD4USQXwaVZ", cids[1])
 	assert.Equal(t, "QmdQEnYhrhgFKPCq5eKc7xb1k7rKyb3fGMitUPKvFAscVK", cids[2])
+}
+
+func TestMustGetChildrenDespiteMissingBlocks(t *testing.T) {
+	f := MustOpenUnixfsCar("./_fixtures/file-3k-and-3-blocks-missing-block.car")
+
+	// {
+	//   "Data": {
+	//     "/": {
+	//       "bytes": "CAIYgBgggAgggAgggAg"
+	//     }
+	//   },
+	//   "Links": [
+	//     {
+	//       "Hash": {
+	//         "/": "QmXakb8wxp4Q9jysbKUDgEnWHXWCg3QEHTaHuhJdDndyN5"
+	//       },
+	//       "Name": "",
+	//       "Tsize": 1035
+	//     },
+	//     {
+	//       "Hash": {
+	//         "/": "Qmed9q9vkn1KDh1NRPTxtUEZvbGouZkkQ5j5oLKtPpNJcf"
+	//       },
+	//       "Name": "",
+	//       "Tsize": 1035
+	//     },
+	//     {
+	//       "Hash": {
+	//         "/": "QmSJX5xgXtnpYnAMbGZQg3YUBwDn75HMpAC6woxqpNgjD4"
+	//       },
+	//       "Name": "",
+	//       "Tsize": 1035
+	//     }
+	//   ]
+	// }
+
+	// The block in the middle was filtered out
+	assert.Equal(t, f.MustGetCid(), "QmZGmS2U9aD3EH8Vtea2fWpurjjry4rCJ2A2dSQFCCSFdB")
+
+	// We should be able to retrieve children CID nonetheless
+	assert.Equal(t, f.MustGetChildrenCids(), []string{
+		"QmXakb8wxp4Q9jysbKUDgEnWHXWCg3QEHTaHuhJdDndyN5",
+		"Qmed9q9vkn1KDh1NRPTxtUEZvbGouZkkQ5j5oLKtPpNJcf",
+		"QmSJX5xgXtnpYnAMbGZQg3YUBwDn75HMpAC6woxqpNgjD4",
+	})
 }